feat: implement power token display
This also adds a new system where the "ACTIVE" string and tokens are both handled as the card's status string, which displays as the button's text. Each card has a method that returns the properly formatted status string. This can be expanded for future text-based card info which changes at runtime.