feat: add two classes in preparation for the stack refactor
This is probably only the first of these commits, since I am planning a pretty ambitious refactor of how card activations work. Basic idea is for each active card, we fill up a stack with "activation triggers" from tokens and the card itself, then process those triggers with a set of observers that look at each item on the stack (the observers will always finish with a 'main' observer that activates the card's main function and then removes it from the stack). Observers are also how we implement passive effects, by adding a new observer for each passive in play.