refactor: implement a very basic form of the event stack
Currently, we have no real observers, just a hardcoded function that executesReally the stack is currently acting more like a queue, but that will probably end up getting refactored. We just need to fill things up in reverse so we can do the FILO behavior that makes a stack a stack.

Also, right now there is a debug function that just checks if an event is from a card with a computer tag as a proof of concept for later observers