refactor: add basic form of observer
This is a very basic computer bonus observer that right now just lives in vars.py. I need to think some about whether observers have a right to be their own class, because right now they are just a wrapper for a function.

Next step will be thinking hard about what I want observers to do and then implementing a card that adds an observer. After that, maybe add a display for observers to the sidebar (presented to users as 'passive effects')