feat: add a passive card that gives bonus points whenever a computer card is activated
This is the first big test of passive cards and works well as a proof of concept.

However, the passive system could still stand to be improved and expanded.

Firstly, right now only active cards (not tokens) are checked for prerun functions (the functions that add observers to the observer list before the card activation loop begins).

Secondly, prerun functions are not handled via the event stack, meaning it is impossible to implement passives which work off of prerun functions.

Thirdly, I believe there are potential bugs/issues which could arise if we ever want to add tokens with passive effects. Right now, all passives are added before the card activation loop starts, meaning there may be issues with creating tokens that need to add observers. Maybe this isn't an issue, and maybe the limitations are actually just an expected emergent behavior? I am not sure, but the topic merits thought. Consider, for example, the issues with adding pollution tokens with negative passive effects that harm the efficacy of card functions.