misc: add energy and half life fields to the Token object
This was the start of a token refactor to allow us to initialize tokens with varying energies and half-lives, but I quickly realized that will be a very long refactor and am pivoting to adding some more passive cards for now. Will probably get back to working on the token refactor next week.