misc: last commit before the big organizational refactor
This commit also serves to add the very start of a documentation file. In the process of making that file I realized that the code could use a decently large organizational refactor in order to keep it clean and extensible--especially since it seems likely that I may be getting some collaborators soon.