refactor: re-organize common.py and rename Observers into Interpreters
The name change is more descriptive—Interpreters are there to interpret StackEvents. "Observer" is already used for some other things and seems to imply that they wait around and watch for something to happen, which is not correct.

The reorganization of  common.py is just a cleanliness thing. Misc variables go at the top, things are organized under comment headings, etc.