refactor: split most gui functionality off into gui.py
There are still a few straggler functions that continue to live in main.py (primarily the functions involved with initializing new item objects). Ideally, these will eventually also be moved over for organizational reasons, but they may need to stay for now. The decision to include the gui button in the item class is making the gui difficult to fully disentangle from the main game logic.