feat: start customizing the layout and theme
This is the start of a series of commits (probably) that bring the theme and layout in line with the vision I mocked up in Photopea. The biggest thing I did here was copy in all the theme-defining folders from minima. Also made it dark mode, sort of. Still got some stuff to work out there, but I am tired and it is time for bed.

 After the theme is sorted out, we should be able to start putting in Paige's card descriptions.