feat: make the card directory page auto-populate with linked card images
This is the meat of the changes to the card directory page. Cards 'know' to add themselves to this page by putting card_page: true in the YAML front matter of their markdown-formatted writeup. I also added some placeholder card pages (currently just copies of the IceCube page) to show what this page looks like when filled out a bit more.

There is still more to do here, like centering the cards and changing the placeholder text, but overall this is pretty good.