feat: improve page styling
This is a rather rough commit containing a bunch of changes, oops. They include:
- Added proper placeholder card pages w/ placeholder images.
- Removed the debug card border
- Replaced the background with one that I have permission to use and is higher-resolution
- Gave the footer a translucent background
- Made the footer display with a full border on card pages

In the future, I could let you filter cards by tag on the card directory page, but right now we just don't have enough for that to make sense. May also need to look into fixing the card display order. That should be easily doable.